requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
pillow
gdown
